Move existing config module to config_utils in preparation for future
