+++ satpy/config_utils.py
