+++ satpy/tests/reader_tests/test_omps_edr.py
@@ -105,7 +105,7 @@ class TestOMPSEDRReader(unittest.TestCase):
 
     def setUp(self):
         """Wrap HDF5 file handler with our own fake handler"""
-        from satpy.config import config_search_paths
+        from satpy.config_utils import config_search_paths
         from satpy.readers.omps_edr import EDRFileHandler, EDREOSFileHandler
         self.reader_configs = config_search_paths(os.path.join('readers', self.yaml_file))
         # http://stackoverflow.com/questions/12219967/how-to-mock-a-base-class-with-python-mock-library
