+++ satpy/tests/test_scene.py
@@ -90,7 +90,7 @@ class TestScene(unittest.TestCase):
 
     def test_init_alone(self):
         from satpy.scene import Scene
-        from satpy.config import PACKAGE_CONFIG_PATH
+        from satpy.config_utils import PACKAGE_CONFIG_PATH
         scn = Scene()
         self.assertEqual(scn.ppp_config_dir, PACKAGE_CONFIG_PATH)
         self.assertFalse(scn.readers, 'Empty scene should not load any readers')
